==> 03_LC02_band_correction.Rmd <==
---
title: "Optical Band Corrections"
author: "Sam Sillen"
date: "2022-03-10"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
# This file uses all the available SR data over wqp/lagos sites (not just in-situ matchups) to create correlation coefficients between different landsat sensors. 

# These coefficients will be applied to 1) the matchup dataset; and 2) the River SR dataset.  

```

# Packages
```{r}
library(tidyverse)
library(lubridate)
library(data.table)
library(broom)
library(feather)
library(ggpubr)

```

# Prepare SR data for joining with in situ 
```{r}

sr_path <- "G:/My Drive/LC02_matchup_pull_algalMask"

sr_raw <- list.files(path=sr_path, pattern=".csv", full.names = T) %>%
  map_df(~ fread(., stringsAsFactors = F))

sr <- sr_raw %>%
 # drop_na(Blue) %>% # Drop all of rows that have been masked but still have lines due to metadata extraction , actually not necessary anymore since pull updates
  mutate(date = date / 1000, # Have to convert miliseconds to seconds for following line 
         datetime =  as.POSIXct(date, origin = "1970-01-01", tz = 'UTC'),
         date = as.POSIXct(date, origin = "1970-01-01", tz = 'UTC'),
         date = as_date(date))

```

```{r}
# Start to build SR file 
sr_clean <- sr %>% 
    filter(pCount_dswe1 > 5, 
    hillShadow > 0, 
    Blue <= 0.2, 
    Blue >= 0,
    Green <= 0.2, 
    Green >=  0, 
    Red <= 0.2, 
    Red >=  0,
    Swir1 <= 0.2, 
    Swir1 >=  0,
    Swir2 <= 0.2, 
    Swir2 >=  0, 
    Nir >=  0, 
    Nir <= 0.2) %>% 
    rename(index = "system:index") %>%
    mutate(LT05 = str_extract(index, pattern = c('LT05')),
           LE07 = str_extract(index, pattern = c('LE07')),
           LC08 = str_extract(index, pattern = c('LC08')),
           LC09 = str_extract(index, pattern = c('LC09')),
           sat = coalesce(LT05, LE07, LC08, LC09)) %>%
    select(-LT05, -LE07, -LC08, -LC09)   %>% 
    mutate(month = as.numeric(month(date)),
      year = year(date),
      hour = hour(datetime)) %>%
    mutate(season = case_when(
      month %in%  9:11 ~ "Fall",
      month %in%  c(12,1,2)  ~ "Winter",
      month %in%  3:5  ~ "Spring",
      TRUE ~ "Summer")) %>%
    mutate(decade= cut(year, breaks = c(1983,1990,1995,2000,2005,2010,2015,2020),
                       labels = c(1990,1995,2000,2005,2010,2015,2020) )) %>%
    rename(landsatID = 'index') %>% 
    dplyr::group_by(SiteID) %>%
    dplyr::mutate(count =n(),
                  max_year=max(year, na.rm=T),
                  min_year = min(year, na.rm=T),
                  n_years = (max_year - min_year)) %>%
    ungroup() 

```

# Coefficients for LS 5
```{r}

sr_clean <- sr_clean  %>%  
    filter(pCount_dswe1 > 5, 
    hillShadow > 0, 
    Blue <= 0.2, 
    Blue >= 0,
    Green <= 0.2, 
    Green >=  0, 
    Red <= 0.2, 
    Red >=  0,
    Swir1 <= 0.2, 
    Swir1 >=  0,
    Swir2 <= 0.2, 
    Swir2 >=  0, 
    Nir >=  0, 
    Nir <= 0.2)

sr_57 <- sr_clean %>%
  filter(sat %in% c("LE07", "LT05")) %>%
  filter(between(date, "1999-01-01", "2012-05-01" )) %>%
  # Filter to site with enough data
  filter(n_years > 10) %>%
  select(SiteID, date, sat, count, n_years, Blue, Red, Green, Nir, Swir1, Swir2) %>%
  gather(Blue:Swir2, Green, key='band', value='value') 

sr_57_rank  <- sr_57 %>%
  droplevels() %>%
  filter(sat =="LT05") %>%
  group_by(band) %>%
  nest() %>%
  mutate( ret = purrr::map(data, ~quantile(.$value, probs = seq(0,1,0.01))),
          ret = purrr::invoke_map(tibble, ret)) %>%
  unnest(ret) %>%
  dplyr::select(-data) %>%
  pivot_longer(
    cols= contains("%")
  ) %>%
  mutate(quant = parse_number(name)/100) %>%
  rename(value_5 = value) %>%
  inner_join(sr_57 %>%
               droplevels() %>%
               filter(sat =="LE07") %>%
               group_by(band) %>%
               nest() %>%
               mutate( ret = purrr::map(data, ~quantile(.$value, probs = seq(0,1,0.01))),
                       ret = purrr::invoke_map(tibble, ret)) %>%
               unnest(ret) %>%
               dplyr::select(-data) %>%
               pivot_longer(
                 cols= contains("%")
               ) %>%
               mutate(quant = parse_number(name)/100) %>%
               rename(value_7 = value) %>%
               dplyr::select(-name),
             by=c("band", "quant")) 

poly_5_trunc <- function(df){
  lm(value_7 ~ poly(value_5, 2, raw=T), data = df %>%
       filter(!quant %in% c(0, 1))  )
}
poly_5_all <- function(df){
  lm(value_7 ~ poly(value_5, 2, raw=T), data = df)
}

# Polynomial correction fit
poly_57 <- sr_57_rank %>%
  ungroup() %>%
#  filter(band != "dw") %>%
  nest(-band) %>%
  mutate( model = purrr::map(data, poly_5_trunc)) %>%
  mutate( model_all = purrr::map(data, poly_5_all)) %>%
  mutate( pred = purrr::map2(model, data, predict)) %>%
  mutate( pred_all = purrr::map2(model_all, data, predict)) %>%
  unnest(c(pred, pred_all, data))  %>%
  dplyr::select(-model, -model_all)

coef_5 <- sr_57_rank %>%
  ungroup() %>%
#  filter(band != "dw") %>%
  filter(!quant %in% c(0, 1)) %>%
  group_by(band)  %>%
  nest() %>%
  mutate( model = purrr::map(data, ~lm(value_7 ~ poly(value_5, 2, raw=T), data = .) %>%
                               tidy %>%
                               dplyr::select(term, estimate) %>%
                               spread(term, estimate))) %>%
  unnest(model) %>%
  dplyr::select(-data) %>%
  rename(band= 1, intercept=2, coef1=3, coef2=4 )  %>%
  mutate(sat = "LT05") %>%
  mutate(fit = "98_quant")
```

```{r}
plot_57 <- ggplot(poly_57) + 
geom_point( aes(x = value_5, y = value_7))+
#geom_point(aes(x=pred_all, y=value_7), color = 'red', alpha =0.5)+
geom_point(aes(x=pred, y = value_7), color = 'red', alpha =0.5)+
geom_abline(aes(slope = 1, intercept =0))+
xlim(0,0.2) +
ylim(0,0.2) +
facet_wrap(~band, scales = "free")+
  labs(x = 'Landsat 5', y = 'Landsat 7') + 
  theme_bw() + 
  theme(panel.grid.major = element_blank(),
            panel.grid.minor = element_blank(), 
           plot.margin = margin(11, 11, 11, 11), axis.text = element_text(colour = 'black', size = 16),strip.text = element_text(size=20), axis.title = element_text(size = 24, colour = 'black'),plot.title = element_text(size = 30, colour = 'black'))

```

# Coefficients for LS 8 
```{r}
sr_78 <- sr_clean %>%
  filter(sat %in% c("LE07", "LC08")) %>%
  filter(date > "2013-04-11" ) %>%
  # filter to site with enough data
  filter(n_years > 10) %>%
  select(SiteID, date, sat, count, n_years, Blue, Red, Green, Nir, Swir1, Swir2) %>%
  gather(Blue:Swir2, key='band', value='value')  

# Do ranking plotting percentiles, joining, and correcting
sr_78_rank  <- sr_78 %>%
  droplevels() %>%
  filter(sat =="LC08") %>%
  group_by(band) %>%
  nest() %>%
  mutate( ret = purrr::map(data, ~quantile(.$value, probs = seq(0,1,0.01))),
          ret = purrr::invoke_map(tibble, ret)) %>%
  unnest(ret) %>%
  dplyr::select(-data) %>%
  pivot_longer(
    cols= contains("%")
  ) %>%
  mutate(quant = parse_number(name)/100) %>%
  rename(value_8 = value) %>%
  inner_join(sr_78 %>%
               droplevels() %>%
               filter(sat =="LE07") %>%
               group_by(band) %>%
               nest() %>%
               mutate( ret = purrr::map(data, ~quantile(.$value, probs = seq(0,1,0.01))),
                       ret = purrr::invoke_map(tibble, ret)) %>%
               unnest(ret) %>%
               dplyr::select(-data) %>%
               pivot_longer(
                 cols= contains("%")
               ) %>%
               mutate(quant = parse_number(name)/100) %>%
               rename(value_7 = value) %>%
               dplyr::select(-name),
             by=c("band", "quant"))  

poly_8_trunc <- function(df){
  lm(value_7 ~ poly(value_8, 2), data = df %>%
       filter(!quant %in% c(0, 1))  )
}
poly_8_all <- function(df){
  lm(value_7 ~ poly(value_8, 2), data = df)
}

poly_78 <- sr_78_rank %>%
  ungroup() %>%
#  filter(band != "dw") %>%
  nest(-band) %>%
  mutate( model = purrr::map(data, poly_8_trunc)) %>%
  mutate( model_all = purrr::map(data, poly_8_all)) %>%
  mutate( pred = purrr::map2(model, data, predict)) %>%
  mutate( pred_all = purrr::map2(model_all, data, predict)) %>%
  unnest(c(pred, pred_all, data)) %>%
  dplyr::select(-model, -model_all)

plot_78 <- ggplot(poly_78) +
  geom_point( aes(x=value_8, y= value_7))+
 # geom_point( aes(x=pred_all, y= value_7), color="red", alpha=0.5)+
  geom_point( aes(x=pred, y= value_7), color="red", alpha=0.5)+
  geom_abline(aes(slope=1, intercept=0)) +
  facet_wrap(~band, scales="free") + 
xlim(0,0.2) +
ylim(0, 0.2) +
  labs(x = 'Landsat 8', y = 'Landsat 7') + 
  theme_bw() + 
  theme(panel.grid.major = element_blank(),
            panel.grid.minor = element_blank(), 
           ,plot.margin = margin(11, 11, 11, 11), axis.text = element_text(colour = 'black', size = 16),strip.text = element_text(size=20), axis.title = element_text(size = 24, colour = 'black'),plot.title = element_text(size = 30, colour = 'black'))

combined_corr <- ggarrange(plot_57, plot_78, nrow = 2)

combined_corr

#ggsave("C:/Users/samsi/OneDrive - University of Pittsburgh/OhioRiver_SR/Plots/sr_band_correction.jpeg", width = 11.5, height = 10, unit = 'in')


coef_8 <- sr_78_rank %>%
  ungroup() %>%
  filter(!quant %in% c(0, 1)) %>%
  group_by(band)  %>%
  nest() %>%
  mutate( model = purrr::map(data, ~lm(value_7 ~ poly(value_8, 2, raw=T), data = .) %>%
                               tidy %>%
                               dplyr::select(term, estimate) %>%
                               spread(term, estimate))) %>%
  unnest(model) %>%
  dplyr::select(-data) %>%
  rename(band= 1, intercept=2, coef1=3, coef2=4 )  %>%
  mutate(sat = "LC08") %>%
  mutate(fit = "98_quant")

coef_7 <- tibble(band = c("Blue", "Red", "Green", "Nir", "Swir1", "Swir2"), intercept = 0, coef1=1, coef2=0, sat= "LE07")

coef_9 <- coef_8 %>% # Landsat 9 sensor is virtually the same as LS 8 , so we are assuming the coefficients should be the same 
mutate(sat = "LC09")

corr_coef <- bind_rows(coef_5, coef_7, coef_8, coef_9) %>%
  ungroup()

#write_feather(corr_coef, "C:/Users/samsi/OneDrive - University of Pittsburgh/LC02_Corr_Coef_algal_mask.feather") # Corr coefficients
#write_feather(sr_clean, "C:/Users/samsi/OneDrive - University of Pittsburgh/LC02_SR_Clean_algal_mask.feather") # Full siteSR with corr data
```
