==> 06_LC02_Build_RiverSR.rmd <==
---
title: "Untitled"
output: github_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Packages 
```{r}
library(tidyverse)
library(lubridate)
library(data.table)
library(feather)
library(ggplot2)
library(httpgd)
library(broom)
library(ggpubr)
```

# Prepare SR data 
```{r}

sr_path <- "G:/My Drive/lc02_polygons_algal_mask/"

sr_raw_v2 <- list.files(path=sr_path, pattern=".csv", full.names = T) %>%
  map_df(~ fread(., stringsAsFactors = F))

binded <- rbind(sr_raw_v1, sr_raw_v2)

sr <- binded %>%
 # drop_na(Blue) %>% # Drop all of rows that have been masked but still have lines due to metadata extraction , actually not necessary anymore since pull updates
  mutate(date = date / 1000) # Have to convert miliseconds to seconds for following line 

# Convert miliseconds since epoch to date
sr$datetime <- as.POSIXct(sr$date, origin = "1970-01-01", tz = 'UTC')
sr$date <- as.POSIXct(sr$date, origin = "1970-01-01", tz = 'UTC')
sr$date <- as_date(sr$date)
```

# Clean SR data 
```{r} 

sr_clean <- sr %>% 
    filter(pCount_algal_mask > 5, 
    hillShadow > 0, 
    Blue <= 1, 
    Blue >= -0.1999725,
    Green <= 1, 
    Green >= -0.1999725, 
    Red <= 1, 
    Red >= -0.1999725,
    Swir1 <= 1, 
    Swir1 >= -0.1999725,
    Swir2 <= 1, 
    Swir2 >= -0.1999725) %>% 
    rename(index = "system:index") %>%
    mutate(LT05 = str_extract(index, pattern = c('LT05')),
           LE07 = str_extract(index, pattern = c('LE07')),
           LC08 = str_extract(index, pattern = c('LC08')),
           LC09 = str_extract(index, pattern = c('LC09')),
           sat = coalesce(LT05, LE07, LC08, LC09)) %>%
    select(-LT05, -LE07, -LC08, -LC09)   %>% 
    mutate(month = as.numeric(month(date)),
      year = year(date),
      hour = hour(datetime)) %>%
    mutate(season = case_when(
      month %in%  9:11 ~ "Fall",
      month %in%  c(12,1,2)  ~ "Winter",
      month %in%  3:5  ~ "Spring",
      TRUE ~ "Summer")) %>%
    mutate(decade= cut(year, breaks = c(1983,1990,1995,2000,2005,2010,2015,2020),
                       labels = c(1990,1995,2000,2005,2010,2015,2020) )) %>%
    rename(landsatID = 'index') %>% 
    ungroup()
```

# Apply correction coefficients
```{r}
corr_coef <- read_feather("C:/Users/samsi/OneDrive - University of Pittsburgh/OhioRiver_SR/LC02_Corr_Coef_algal_mask.feather") %>%
  ungroup()

correction_coef_99 <- corr_coef %>%
  filter(fit %in% c(NA, "98_quant")) %>%
  dplyr::select(-fit)

sr_clean <- sr_clean %>%
  mutate(rn = row_number()) %>%
  mutate(sat = as.character(sat)) %>%
  gather(Red ,Green, Blue, Nir, Swir1, Swir2, key='band', value='value') %>%
  group_by(band, sat) %>%
  left_join(correction_coef_99, by=c("band", "sat")) %>%
  mutate(value_cor=  coef2*value^ 2 + coef1*value + intercept) %>%
  ungroup() %>%
  mutate(value_cor = ifelse(value_cor <=0, value, value_cor)) %>%
  dplyr::select(-intercept, -coef1, -coef2) %>%
  pivot_wider(names_from = band,
              values_from = c("value", "value_cor"))  %>%
    rename_at(vars(starts_with("value_")),
           function(x) stringr::str_replace_all(x, "value_", "")) %>%
    rename_at(vars(Red, Green, Blue, Nir, Swir1, Swir2),function(x) paste0(x,"_raw")) %>%
    rename_at(vars(starts_with("cor")),            funs(stringr::str_replace_all(., "cor_", ""))) 

```


# Add band ratios, color metrics 
```{r}
chroma <- function(R, G, B) {
  require(colorscience)
  require(tidyverse) 

# Converst R,G, and B spectral reflectance to dominant wavelength based
# on CIE chromaticity color space

# see Wang et al 2015. MODIS-Based Radiometric Color Extraction and
# Classification of Inland Water With the Forel-Ule
# Scale: A Case Study of Lake Taihu

# chromaticity.diagram.color.fill()
Xi <- 2.7689*R + 1.7517*G + 1.1302*B
Yi <- 1.0000*R + 4.5907*G + 0.0601*B
Zi <- 0.0565*G + 5.5943*B

x <-  Xi / (Xi + Yi +  Zi)
y <-  Yi / (Xi + Yi +  Zi)
z <-  Zi / (Xi + Yi +  Zi)

# Calculate hue angle
alpha <- atan2( (x - (1/3)), (y - (1/3))) * 180/pi

# Make look up table for hue angle to wavelength conversion
cie <- cccie31 %>%
  dplyr::mutate(a = atan2( (x - (1/3)), (y - (1/3))) * 180/pi) %>%
  dplyr::filter(wlnm <= 700) %>%
  dplyr::filter(wlnm >=380) 

# Find nearest dominant wavelength to hue angle
wl <- cie[as.vector(sapply(alpha,function(x) which.min(abs(x - cie$a)))), 'wlnm']

return(wl)
}


pull_transform <- function(df, maxRGB=1, RGB=F) {
  
  if(RGB == T) { 
    
    maxRGB <- maxRGB
    
    data <- df %>%
      filter_at(vars(Red, Green, Blue), all_vars(.< maxRGB))
    
  }else{ 
    data <- df    

    maxRGB <- df  %>%
      dplyr::select(Red, Green, Blue) %>%
      dplyr::summarise(maxRGB = max(., na.rm=F)) 
    
    maxRGB <- maxRGB$maxRGB
  }
  
  data <- data %>%
    dplyr::mutate(NR = Nir/Red,
                  BR = Blue/Red,
                  GR = Green/Red,
                  SR = Swir1/Red,
                  BG = Blue/Green,
                  RG = Red/Green, 
                  NG = Nir/Green,
                  SG = Swir1/Green,
                  BN = Blue/Nir,
                  GN = Green/Nir,
                  RN = Red/Nir,
                  SN = Swir1/Nir,
                  BS = Blue/Swir1,
                  GS = Green/Swir1,
                  RS = Red/Swir1,
                  NS = Nir/Swir1,
                  R.GN = Red/ (Green + Nir),
                  R.GB = Red/ (Green + Blue),
                  R.GS = Red/ (Green + Swir1),
                  R.BN = Red/ (Blue + Nir),
                  R.BS = Red/ (Blue + Swir1),
                  R.NS = Red/ (Nir + Swir1),
                  G.BR = Green/ (Blue + Red),
                  G.BN = Green / (Blue + Nir),
                  G.BS = Green / (Blue + Swir1),
                  G.RN = Green / (Red + Nir),
                  G.RB = Green / (Red + Blue),
                  G.NS = Green / (Nir + Swir1),
                  B.RG = Blue / (Red + Green),
                  B.RN = Blue / (Red + Nir),
                  B.RS = Blue / (Red + Swir1),
                  B.GN = Blue / (Green + Nir),
                  B.GS = Blue / (Green + Swir1),
                  B.NS = Blue / (Nir + Swir1),
                  N.RG = Nir / (Red + Green),
                  N.RB = Nir / (Red + Blue),
                  N.RS = Nir / (Red + Swir1),
                  N.GB = Nir / (Green + Blue),
                  N.GS = Nir / (Green + Swir1),
                  N.BS = Nir / (Blue  + Swir1),
                  GR2 = (Green + Red) / 2,
                  GN2 = (Green + Nir) / 2,
                  # Adding bloom metrics 
                  BR_G = (Blue - Red) / Green,
                  NS_NR = (Nir - Swir1) / (Red - Swir1),
                  fai = Nir - (Red + (Swir1-Red)*((830-660)/(1650-660))),
                  GCI = Nir/(Green-1),
                  IRG = Red-Green,
                  SABI = (Nir-Red)/(Blue + Green),
                  KIVU = (Blue-Red)/Green,
                  GB = Green/Blue,
                  GNDVI = (Nir-Green)/(Nir+Green),
                  EVI = 2.5*((Nir-Red)/(Nir+((6*Red)-(7.5*Blue))+1)),
                  KAB = 1.67-3.94*log(Blue)+3.78*log(Red),
                  KRL = (((Blue/Red)*Nir)-98)/0.75,
                  N_S = Nir - Swir1,
                  N_R = Nir - Red,
                  ndvi = ((Nir-Red)/(Nir+Red)),
                  ndwi = ((Green- Swir1)/(Green + Swir1)),
                  ndssi = ((Blue - Nir)/ (Blue + Nir)),
                  gn.gn= ((Green- Nir)/ (Green + Nir)))

  color <- data %>% filter(Blue > 0 & Green > 0 & Red > 0) %>% 
    dplyr:: mutate(hue = rgb2hsv(r=Red, g=Green, b=Blue, maxColorValue = maxRGB)[1,],
            saturation = rgb2hsv(r=Red, g=Green,  b=Blue, maxColorValue = maxRGB)[2,],
            bright = rgb2hsv(r=Red, g=Green,  b=Blue, maxColorValue = maxRGB)[3,],
            bright_tot = (Red + Green + Nir +Blue),
            dw = chroma(R=Red, G=Green, B=Blue),
            hexcolor = rgb(r=Red, g=Green, b=Blue, maxColorValue = maxRGB))

  color_na <- data %>% filter(!rn %in% color$rn) %>% 
    dplyr:: mutate(hue = 'NA',
            saturation = 'NA',
            bright = 'NA',
            bright_tot = 'NA',
            dw = 'NA', 
            hexcolor = 'NA')

  data <-rbind(color, color_na) 

  return(data)
}

LC02_Polygons <- pull_transform(sr_clean, RGB=F, maxRGB=1) 

LC02_Polygons$dw <-  as.numeric(LC02_Polygons$dw)

LC02_Polygons$saturation = as.numeric(data$saturation)

LC02_Polygons$hue <- as.numeric(LC02_Polygons$hue)

write_feather(LC02_Polygons, "C:/Users/samsi/OneDrive - University of Pittsburgh/OhioRiver_SR/Data/full_lc02_polygons_algal_mask.feather")

```

# Build corr figure (time series)
```{r}
library(ggplot2)
library(ggthemes)

river_sr <- read_feather("/Users/samsillen/OneDrive - University of Pittsburgh/OhioRiver_SR/Data/LC02_Polygons_COMID_algal_mask_orb.feather")

site <- river_sr %>% filter(comid == 3785895) %>% 
mutate(sat = fct_relevel(sat, 
           'LT05', 'LE07','LC08'))
site$dw
corr_blue <- ggplot(site %>% filter(Blue > 0), aes(y = Blue, x = date, colour = sat)) +
                              geom_point() + 
                              geom_line(lwd = 0.7, alpha = 0.8)+ 
                              scale_color_manual(values=c( "#c6dbef",'#8EA8D6','#3975B9', '#20436D')) +
   #   xlab("Year") +
      ylab("SR Corrected Blue")+
      theme_few() +
      theme(axis.text = element_text(colour = 'black', size = 20), axis.title = element_text(size = 20, colour = 'black'),plot.title = element_text(size = 30, colour = 'black'),
            # plot.background = element_rect(fill = '#292b30', color = 'NA'),
            # panel.grid.major = element_blank(),
            #  panel.grid.minor = element_blank(), 
            #panel.border = element_blank(),
            #panel.background = element_rect(fill = '#292b30'),
            #legend.background = element_rect(fill = '#292b30'), 
            legend.position = 'none', axis.title.x = element_blank(), plot.margin = unit(c(2.5,0.5,0.5,0.5), "cm"))

raw_blue <- ggplot(site %>% filter(Blue > 0),aes(y = Blue_raw, x = date, colour = sat)) +
                              geom_point() + 
                              geom_line(lwd = 0.7, alpha = 0.8)+ 
                              scale_color_manual(values=c( "#c6dbef",'#8EA8D6','#3975B9', '#20436D')) +
   #   xlab("Year") +
      ylab("SR Raw Blue")+
      theme_few() +
      theme(axis.text = element_text(colour = 'black', size = 20), axis.title = element_text(size = 20, colour = 'black'),plot.title = element_text(size = 30, colour = 'black'),
            # plot.background = element_rect(fill = '#292b30', color = 'NA'),
            # panel.grid.major = element_blank(),
            #  panel.grid.minor = element_blank(), 
            #panel.border = element_blank(),
            #panel.background = element_rect(fill = '#292b30'),
            #legend.background = element_rect(fill = '#292b30'), 
            legend.position = 'none', axis.title.x = element_blank(), plot.margin = unit(c(2.5,0.5,0.5,0.5), "cm"))

corr_green <- ggplot(site %>% filter(Green > 0), aes(y = Green, x = date, colour = sat)) +
                              geom_point() + 
                              geom_line(lwd = 0.7, alpha = 0.8)+ 
                              scale_color_manual(values=c( "#c7e9c0",'#74c476','#238b45', '#00441b')) +
   #   xlab("Year") +
      ylab("SR Corrected Green")+
      theme_few() +
      theme(axis.text = element_text(colour = 'black', size = 20), axis.title = element_text(size = 20, colour = 'black'),plot.title = element_text(size = 30, colour = 'black'),
            # plot.background = element_rect(fill = '#292b30', color = 'NA'),
            # panel.grid.major = element_blank(),
            #  panel.grid.minor = element_blank(), 
            #panel.border = element_blank(),
            #panel.background = element_rect(fill = '#292b30'),
            #legend.background = element_rect(fill = '#292b30'), 
            legend.position = 'none', axis.title.x = element_blank(), plot.margin = unit(c(2.5,0.5,0.5,0.5), "cm"))

raw_green <- ggplot(site %>% filter(Green > 0),aes(y = Green_raw, x = date, colour = sat)) +
                              geom_point() + 
                              geom_line(lwd = 0.7, alpha = 0.8)+ 
                              scale_color_manual(values=c( "#c7e9c0",'#74c476','#238b45', '#00441b')) +
   #   xlab("Year") +
      ylab("SR Raw Green")+
      theme_few() +
      theme(axis.text = element_text(colour = 'black', size = 20), axis.title = element_text(size = 20, colour = 'black'),plot.title = element_text(size = 30, colour = 'black'),
            # plot.background = element_rect(fill = '#292b30', color = 'NA'),
            # panel.grid.major = element_blank(),
            #  panel.grid.minor = element_blank(), 
            #panel.border = element_blank(),
            #panel.background = element_rect(fill = '#292b30'),
            #legend.background = element_rect(fill = '#292b30'), 
            legend.position = 'none', axis.title.x = element_blank(), plot.margin = unit(c(2.5,0.5,0.5,0.5), "cm"))

plot <- ggarrange(raw_blue, corr_blue, raw_green, corr_green, nrow = 2, ncol = 2)
plot

ggsave("/Users/samsillen/OneDrive - University of Pittsburgh/OhioRiver_SR/Plots/correction_timerseries.tiff", width = 12, height = 12.5, unit = 'in')


```
